+++ BE/src/main/java/com/team09/issue_tracker/login/interceptor/InterceptorConfiguration.java
@@ -1,0 +1,21 @@
+package com.team09.issue_tracker.login.interceptor;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@RequiredArgsConstructor
+@Configuration
+public class InterceptorConfiguration implements WebMvcConfigurer {
+
+	private final HandlerInterceptor loginInterceptor;
+
+	@Override
+	public void addInterceptors(InterceptorRegistry registry) {
+		registry.addInterceptor(loginInterceptor)
+			.addPathPatterns("/**")
+			.excludePathPatterns("/login/**");
+	}
+}
